FT2-46: Removed unnecessary file.
