+++ app/controller/login_control/logout.php
