+++ app/controller/login_control/otp.php
@@ -1,5 +1,6 @@
 <?php
-require_once 'email_process.php';
+
+require_once __DIR__. '/email_process.php';
 
 session_start();
 $mail = new EmailProcess();
