+++ app/controller/login_control/signup_controller.php
@@ -1,6 +1,6 @@
 <?php
 
-require './model/query.php';
+require_once __DIR__. '/../../model/query.php';
 
 session_start();
 if (isset($_SESSION["data"])) {
