+++ app/controller/posts_control/new_password_controller.php
