+++ app/core/cred_example.php
@@ -6,3 +6,5 @@ $servername = '';
 $username = '';
 $db_password = '';
 $dbname = '';
+$client_id = "";
+$client_secret = "";
