+++ app/model/query.php
@@ -1,5 +1,5 @@
 <?php
-require_once("database_connection.php");
+require_once(__DIR__. "/database_connection.php");
 
 class Queries
 {
