+++ app/view/css/home_style.css
@@ -6,7 +6,7 @@
 body {
   margin: 0;
   font-family: Arial, sans-serif;
-  background-color: #ced4da;
+  background-color: #c9dbef;
 }
 .container {
   width: 30%;
